third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbfactory_open5-expected.txt

This is a testharness.js-based test.
PASS IDBFactory.open() - new database is empty 
Harness: the test ran to completion.
